# Detect CPU model name regardless of system language