--- a/run.sh
+++ b/run.sh
@@ -51,7 +51,7 @@ arch=$(uname -m)
 if [[ "${os}" == "darwin"* || "${os}" == "freebsd"* ]]; then
     model=$(sysctl -n machdep.cpu.brand_string)
 elif [[ "${os}" == "linux-gnu"* ]]; then
-    model=$(lscpu | grep "Model name" | awk -F: '{print $2}' | sed -e 's/^[[:space:]]*//')
+    model=$(lscpu --extended=MODELNAME | awk -F: 'NR==2 {print $1; exit}')
 else
     model="Unknown"
 fi